+++ bpdm-gate-api/src/main/kotlin/org/eclipse/tractusx/bpdm/gate/api/model/SharableRelationType.kt
@@ -19,12 +19,6 @@
 
 package org.eclipse.tractusx.bpdm.gate.api.model
 
-import java.time.Instant
-
-data class RelationOutputDto(
-    val externalId: String,
-    val relationType: SharableRelationType,
-    val sourceBpnL: String,
-    val targetBpnL: String,
-    val updatedAt: Instant
-)
+enum class SharableRelationType {
+    IsAlternativeHeadquarterFor
+}