+++ bpdm-gate/src/main/kotlin/org/eclipse/tractusx/bpdm/gate/service/BusinessPartnerMappings.kt
@@ -159,7 +159,8 @@ class BusinessPartnerMappings {
                 entity.sharingState.externalId,
                 "Missing address confidence criteria"
             ),
-            states = toStateDtos(entity.states, BusinessPartnerType.LEGAL_ENTITY)
+            states = toStateDtos(entity.states, BusinessPartnerType.LEGAL_ENTITY),
+            relations = entity.relations.map { BusinessPartnerOutputRelationDto(it.relationType, it.sourceBpnL, it.targetBpnL) }
         )
     }