+++ bpdm-gate/src/main/kotlin/org/eclipse/tractusx/bpdm/gate/service/OutputUpsertMappings.kt
@@ -24,11 +24,21 @@ import org.eclipse.tractusx.bpdm.gate.entity.*
 import org.eclipse.tractusx.bpdm.gate.entity.generic.*
 import org.eclipse.tractusx.bpdm.gate.model.upsert.output.*
 import org.springframework.stereotype.Service
+import java.time.Instant
+import java.util.*
 
 @Service
 class OutputUpsertMappings {
 
-    fun toEntity(upsertData: OutputUpsertData, sharingState: SharingStateDb): BusinessPartnerDb {
+    fun toEntity(upsertData: OutputUpsertDataWithRelations, sharingState: SharingStateDb): BusinessPartnerDb {
+        return toEntity(
+            upsertData = upsertData.businessPartnerUpsertData,
+            sharingState = sharingState,
+            relations = upsertData.relationUpsertData.map { RelationOutputDb(it.relationType, it.sourceBpnL, it.targetBpnL, Instant.now()) }.toSortedSet()
+        )
+    }
+
+    fun toEntity(upsertData: OutputUpsertData, sharingState: SharingStateDb, relations: SortedSet<RelationOutputDb>): BusinessPartnerDb {
         return with(upsertData) {
             BusinessPartnerDb(
                 sharingState = sharingState,
@@ -50,6 +60,7 @@ class OutputUpsertMappings {
                 legalEntityConfidence = legalEntityConfidence.toEntity(),
                 siteConfidence = siteConfidence?.toEntity(),
                 addressConfidence = addressConfidence.toEntity(),
+                relations = relations.map { RelationOutputDb(it.relationType, it.sourceBpnL, it.targetBpnL, Instant.now()) }.toSortedSet()
             )
         }
     }